LabelsAPIWrapper
