+++ packages/components/src/internal/components/labelsPrinting/BarTenderSettingsForm.spec.tsx
