+++ packages/components/src/internal/components/labelsPrinting/LabelsConfigurationPanel.spec.tsx
