+++ packages/components/src/internal/components/labelsPrinting/constants.ts
