+++ packages/components/src/internal/components/labelsPrinting/utils.ts
