+++ packages/components/src/internal/components/project/ProjectSettings.tsx
@@ -10,7 +10,7 @@ import { Container } from '../base/models/Container';
 
 import { useAdminAppContext } from '../administration/useAdminAppContext';
 
-import { BarTenderSettingsForm } from '../labels/BarTenderSettingsForm';
+import { BarTenderSettingsForm } from '../labelsPrinting/BarTenderSettingsForm';
 import { NameIdSettings } from '../settings/NameIdSettings';
 import { biologicsIsPrimaryApp, isAppHomeFolder, isProtectedDataEnabled } from '../../app/utils';
 import { ProtectedDataSettingsPanel } from '../administration/ProtectedDataSettingsPanel';
