+++ packages/components/src/internal/test/testHelpers.tsx
@@ -13,7 +13,7 @@ import { ServerContext, ServerContextProvider } from '../components/base/ServerC
 import {
     LabelPrintingContextProps,
     LabelPrintingContextProvider,
-} from '../components/labels/LabelPrintingContextProvider';
+} from '../components/labelsPrinting/LabelPrintingContextProvider';
 import { GlobalStateContextProvider } from '../GlobalStateContext';
 import { URL_MAPPERS, URLService } from '../url/URLResolver';
